refactor: changed structure of Game, View
